requirements.txt
mlflow
scikit-learn
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
flask
flask-Cors